Refactor: Update existing components instead of creating enhanced versions
- Merged all EnhancedBaseBackground features into BaseBackground
- Merged all EnhancedImagesCarousel features into ImagesCarousel
- Removed enhanced component files
- Updated all imports to use original component names

This simplifies the codebase by updating existing components directly
rather than maintaining separate enhanced versions. All v4 features
(Google Fonts, motion blur, noise texture) are now integrated into
the original components.
