+++ src/components/ImagesCarousel.tsx
@@ -1,8 +1,8 @@
 import { spring, useCurrentFrame, useVideoConfig } from "remotion";
-
+import { CameraMotionBlur } from "@remotion/motion-blur";
 import { Image } from "./Image";
 
-export const ImagesCarousel = ({ images }) => {
+export const ImagesCarousel = ({ images }: { images: string[] }) => {
   const frame = useCurrentFrame();
   const { fps } = useVideoConfig();
 
@@ -27,22 +27,30 @@ export const ImagesCarousel = ({ images }) => {
     config: springConfig,
   });
 
+  const currentTranslate = frame < 110 ? translate1 : translate2;
+  // Calculate motion blur intensity based on transition state
+  const isTransitioning =
+    (frame > 45 && frame < 65) || (frame > 105 && frame < 125); // Wider transition window
+  const shutterAngle = isTransitioning ? 270 : 0.1; // Higher angle for more noticeable blur
+
   return (
-    <div className="relative aspect-[16/9] overflow-hidden">
-      <div
-        className="flex h-full w-full"
-        style={{
-          transform: `translateX(-${frame < 110 ? translate1 : translate2}%)`,
-        }}
-      >
-        {images.map((image) => (
-          <Image
-            src={image}
-            className="h-full w-full flex-shrink-0 object-contain object-center"
-            key={image}
-          />
-        ))}
-      </div>
+    <div className="relative aspect-[16/9] overflow-hidden rounded-xl">
+      <CameraMotionBlur shutterAngle={shutterAngle} samples={8}>
+        <div
+          className="flex h-full w-full"
+          style={{
+            transform: `translateX(-${currentTranslate}%)`,
+          }}
+        >
+          {images.map((image) => (
+            <Image
+              src={image}
+              className="h-full w-full flex-shrink-0 object-contain object-center"
+              key={image}
+            />
+          ))}
+        </div>
+      </CameraMotionBlur>
     </div>
   );
 };
